requirements/local.txt
-r base.txt

psycopg2-binary==2.9.5  # https://github.com/psycopg/psycopg2

# Code quality
# ------------------------------------------------------------------------------
black==22.12.0  # https://github.com/psf/black

# Django Extensions
django-extensions==3.1.3  # https://github.com/django-extensions/django-extensions
djangorestframework==3.14.0
